news_assistant_ready/requirements.txt
python-telegram-bot==20.3
feedparser
pandas
psutil
